preparation_dataset/pyproject.toml
[tool.poetry]
name = "remodel-unox-preparation-dataset"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
google-cloud-bigquery = "^3.34.0"
pandas = "^2.3.0"
python-dotenv = "^1.1.1"
db-dtypes = "^1.4.3"
ipython = "^9.3.0"
geopandas = "^1.1.1"
osmnx = "^2.0.4"
tqdm = "^4.67.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"



[tool.poetry.scripts]
main = "main:main"